Add backface of mesh by default. Closes https://github.com/deathcap/voxel-webview/issues/11
Previously the mesh only had one side, so it was not visible from the
rear, or was, depending on backface visibility options (it was visible
in the demo, but not voxel-webview). Add opts.backface (default true) to
add two other reverse-wound triangles to the mesh to always render the back.